1.修复首次使用授权后界面不刷新问题   2.解决部分内存暴增问题

+++ XMNPhotoPickerFramework/Models/XMNAlbumModel.m
@@ -67,7 +67,7 @@
 #pragma mark - Getters
 
 - (NSString *)description {
-    return [NSString stringWithFormat:@"\n-----album desc start--\nalbum :%@ have %ld photos \nresult :%@\n-----album desc end----",self.name,self.count,self.fetchResult];
+    return [NSString stringWithFormat:@"\n-----album desc start--\nalbum :%@ have %ld photos \nresult :%@\n-----album desc end----",self.name,(unsigned long)self.count,self.fetchResult];
 }
